Move google login code to a different file
