+++ src/gapi.js
@@ -1,6 +1,25 @@
-/* eslint-disable import/prefer-default-export */
+import { CLIENT_ID, API_KEY } from './config'
+
 export const gapi = window.gapi
-export const SCOPES = 'https://www.googleapis.com/auth/calendar'
-export const DISCOVERY_DOCS = [
+const SCOPES = 'https://www.googleapis.com/auth/calendar'
+const DISCOVERY_DOCS = [
   'https://www.googleapis.com/discovery/v1/apis/calendar/v3/rest'
 ]
+
+export const handleClientLoad = () => {
+  gapi.load('client:auth2', initClient)
+}
+const initClient = () => {
+  gapi.client
+    .init({
+      apiKey: API_KEY,
+      clientId: CLIENT_ID,
+      discoveryDocs: DISCOVERY_DOCS,
+      scope: SCOPES
+    })
+    .then(
+      function () {
+        gapi.auth2.getAuthInstance().signIn()
+      }
+    )
+}
